Separate the selection blueprint logic into individual class.
